Move code into NexusConverter
Refs #268
